+++ src/components/Home/WalkInfo/fallback.tsx
