+++ src/components/Home/WalkInfo/loader.tsx
